# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AcceptedOut": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "job": {
            "title": "Job",
            "type": "string"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          }
        },
        "required": [
          "accepted",
          "job"
        ],
        "title": "AcceptedOut",
        "type": "object"
      },
      "Datum": {
        "description": "One value with its provenance, or one explicit gap.\n\nA gap is {\"value\": null, \"is_gap\": true, \"note\": ..., \"gap_reason\": ...}. Every key is present\nin both cases.",
        "properties": {
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 4217, for money values",
            "title": "Currency"
          },
          "fiscal_year": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "e.g. \"2024\" or \"2024/25\"",
            "title": "Fiscal Year"
          },
          "gap_reason": {
            "anyOf": [
              {
                "enum": [
                  "not_searched",
                  "searched_not_found",
                  "source_unreachable",
                  "not_public"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "present exactly when value is null",
            "title": "Gap Reason"
          },
          "is_gap": {
            "description": "true exactly when value is null",
            "title": "Is Gap",
            "type": "boolean"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "present whenever value is null",
            "title": "Note"
          },
          "quote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "verbatim, at most 40 words",
            "title": "Quote"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "scope": {
            "anyOf": [
              {
                "enum": [
                  "global",
                  "nepal_only",
                  "unknown"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "whether a figure is global or Nepal only",
            "title": "Scope"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "present whenever value is not null",
            "title": "Source Url"
          },
          "value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "description": "null means no value was found; see gap_reason",
            "title": "Value"
          },
          "value_type": {
            "anyOf": [
              {
                "enum": [
                  "string",
                  "integer",
                  "number",
                  "boolean",
                  "money",
                  "date"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Type"
          },
          "verification": {
            "enum": [
              "self_reported",
              "register_confirmed",
              "externally_audited",
              "third_party_reported",
              "unverified"
            ],
            "title": "Verification",
            "type": "string"
          }
        },
        "required": [
          "is_gap",
          "verification"
        ],
        "title": "Datum",
        "type": "object"
      },
      "DatumHistoryEntry": {
        "properties": {
          "datum": {
            "$ref": "#/components/schemas/Datum"
          },
          "superseded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded At"
          },
          "valid_from": {
            "format": "date-time",
            "title": "Valid From",
            "type": "string"
          }
        },
        "required": [
          "datum",
          "valid_from"
        ],
        "title": "DatumHistoryEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OrgDetail": {
        "properties": {
          "aliases": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "data": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Datum"
            },
            "description": "every datum by its JSON path, gaps included, e.g. financial_transparency.income",
            "title": "Data",
            "type": "object"
          },
          "data_gaps": {
            "description": "paths that stayed empty after a real search",
            "items": {
              "type": "string"
            },
            "title": "Data Gaps",
            "type": "array"
          },
          "hq_city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hq City"
          },
          "hq_country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hq Country"
          },
          "last_updated": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "name_common": {
            "title": "Name Common",
            "type": "string"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "org_type": {
            "title": "Org Type",
            "type": "string"
          },
          "registrations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RegistrationOut"
            },
            "title": "Registrations",
            "type": "array"
          },
          "research_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Notes"
          },
          "statements": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StatementOut"
            },
            "title": "Statements",
            "type": "array"
          },
          "warnings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WarningOut"
            },
            "title": "Warnings",
            "type": "array"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "org_id",
          "name_common",
          "org_type"
        ],
        "title": "OrgDetail",
        "type": "object"
      },
...
      "RunOut": {
        "properties": {
          "cost_usd": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "git_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job": {
            "title": "Job",
            "type": "string"
          },
          "rows_rejected": {
            "title": "Rows Rejected",
            "type": "integer"
          },
          "rows_skipped": {
            "title": "Rows Skipped",
            "type": "integer"
          },
          "rows_written": {
            "title": "Rows Written",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "job",
          "status",
          "started_at",
          "rows_written",
          "rows_skipped",
          "rows_rejected"
        ],
        "title": "RunOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WarningOut": {
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "occurred_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred On"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "source_url",
          "note"
        ],
        "title": "WarningOut",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Who is responding to a disaster, where, since when, and from which source. Every value carries its provenance; a missing value is an explicit, visible state. This API does not rate, score or rank organisations.",
    "title": "Spenden-Transparenz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/ingest/{job}": {
      "post": {
        "operationId": "trigger_ingest_v1_admin_ingest__job__post",
        "parameters": [
          {
            "in": "path",
            "name": "job",
            "required": true,
            "schema": {
              "title": "Job",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger one ingestion job.",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/runs": {
      "get": {
        "operationId": "list_runs_v1_admin_runs_get",
        "parameters": [
          {
            "description": "page size, at most 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "page size, at most 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-admin-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "title": "Response List Runs V1 Admin Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent ingestion runs.",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/orgs": {
      "get": {
        "operationId": "list_orgs_v1_orgs_get",
        "parameters": [
          {
            "description": "repeatable",
            "in": "query",
            "name": "org_type",
            "required": false,
            "schema": {
              "default": [],
              "description": "repeatable",
              "items": {
                "type": "string"
              },
              "title": "Org Type",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "hq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "local",
                    "international"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hq"
            }
          },
          {
            "description": "name search, matched case-insensitively",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name search, matched case-insensitively",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "enum": [
                "latest",
                "name",
                "least_data"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "description": "page size, at most 100",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "page size, at most 100",
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrgDetail"
                  },
                  "title": "Response List Orgs V1 Orgs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Organisations.",
        "tags": [
          "organisations"
        ]
      }
    },
    "/v1/orgs/{org_id}": {
      "get": {
        "operationId": "get_org_v1_orgs__org_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One organisation, gaps included.",
        "tags": [
          "organisations"
        ]
      }
    },
    "/v1/orgs/{org_id}/history": {
      "get": {
        "operationId": "get_org_history_v1_orgs__org_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "org_id",
            "required": true,
            "schema": {
              "title": "Org Id",
              "type": "string"
            }
          },
          {
            "description": "datum path, e.g. financial_transparency.income",
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "description": "datum path, e.g. financial_transparency.income",
              "maxLength": 200,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DatumHistoryEntry"
                  },
                  "title": "Response Get Org History V1 Orgs  Org Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "How one value changed.",
        "tags": [
          "organisations"
        ]
      }
    },
...
  }
}